+++ CodeFirst/Models/ViewModels/OrderWithUserViewModel.cs
@@ -27,9 +27,9 @@ namespace CodeFirst.Models.ViewModels
 
         public int Quantity { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public double Total { get; set; }
+        public decimal Total { get; set; }
 
         public virtual ApplicationUser ApplicationUser { get; set; }
     }
